[Update]いいね機能の実装

+++ app/controllers/favorites_controller.rb
@@ -1,13 +1,17 @@
 class FavoritesController < ApplicationController
   def create
     @post = Post.find(params[:post_id])
-    favorite = current_user.favorites.new(post_id: @post.id)
-    favorite.save
+    unless current_user == @post.user
+      favorite = current_user.favorites.new(post_id: @post.id)
+      favorite.save
+    end
   end
 
   def destroy
     @post = Post.find(params[:post_id])
-    favorite = current_user.favorites.find_by(post_id: @post.id)
-    favorite.destroy
+    unless current_user == @post.user
+      favorite = current_user.favorites.find_by(post_id: @post.id)
+      favorite.destroy
+    end
   end
 end
